Add files via upload
Work-in-Progress game file which can run in terminal. I do think that we should divide the work among ourselves. However, this is just some code I wrote the other day. This code only works for a specific input and there are a lot of user stories/cases that I have not accounted for yet. The input you need to give it after running the file is as follows: TIGER -> 4,3 -> 1 and it should update and print the new grid. Also, apologies for posting it late.